requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
openai
python-dotenv
python-docx
pandas
pdfplumber
python-pptx
msal
azure-identity
azure-storage-blob
PyJWT
openpyxl